+++ src/io/stdin.rs
@@ -5,7 +5,7 @@ use crate::{
         slice::{NSTDSlice, NSTDSliceMut},
         str::nstd_core_str_from_bytes_unchecked,
     },
-    io::NSTDIOError,
+    io::{NSTDIOError, NSTDIOResult},
     string::{nstd_string_push_str, NSTDString},
     vec::NSTDVec,
     NSTDUInt,
@@ -42,11 +42,10 @@ pub fn nstd_io_stdin() -> NSTDStdin {
 ///
 /// - `NSTDSliceMut *buffer` - The buffer to fill with data from stdin.
 ///
-/// - `NSTDUInt *read` - Returns as the number of bytes read from stdin.
-///
 /// # Returns
 ///
-/// `NSTDIOError errc` - The I/O operation error code.
+/// `NSTDIOResult read` - The number of bytes read from `handle` on success, or the I/O operation
+/// error code on failure.
 ///
 /// # Safety
 ///
@@ -56,20 +55,11 @@ pub fn nstd_io_stdin() -> NSTDStdin {
 pub unsafe fn nstd_io_stdin_read(
     handle: &mut NSTDStdin,
     buffer: &mut NSTDSliceMut,
-    read: &mut NSTDUInt,
-) -> NSTDIOError {
+) -> NSTDIOResult {
     #[cfg(not(unix))]
-    {
-        let (err, r) = crate::io::stdio::read(handle, buffer);
-        *read = r;
-        err
-    }
+    return crate::io::stdio::read(handle, buffer);
     #[cfg(unix)]
-    {
-        let (err, r) = crate::os::unix::io::stdio::read(handle.lock().as_raw_fd(), buffer);
-        *read = r;
-        err.into()
-    }
+    return crate::os::unix::io::stdio::read(handle.lock().as_raw_fd(), buffer).into();
 }
 
 /// Continuously reads data from stdin into a buffer until EOF is reached.
@@ -77,7 +67,7 @@ pub unsafe fn nstd_io_stdin_read(
 /// # Note
 ///
 /// If extending the buffer fails, an error code of `NSTD_IO_ERROR_OUT_OF_MEMORY` will be returned.
-/// This does not mean `read` will return as 0 in this case.
+/// This does not mean there were no bytes read from `handle` in this case.
 ///
 /// # Parameters:
 ///
@@ -85,36 +75,23 @@ pub unsafe fn nstd_io_stdin_read(
 ///
 /// - `NSTDVec *buffer` - The buffer to be extended with data from stdin.
 ///
-/// - `NSTDUInt *read` - Returns as the number of bytes read from stdin.
-///
 /// # Returns
 ///
-/// `NSTDIOError errc` - The I/O operation error code.
+/// `NSTDIOResult read` - The number of bytes read from `handle` on success, or the I/O operation
+/// error code on failure.
 ///
 /// # Panics
 ///
 /// This function will panic if `buffer`'s length in bytes ends up exceeding `NSTDInt`'s max value.
 #[inline]
 #[nstdapi]
-pub fn nstd_io_stdin_read_all(
-    handle: &mut NSTDStdin,
-    buffer: &mut NSTDVec,
-    read: &mut NSTDUInt,
-) -> NSTDIOError {
+pub fn nstd_io_stdin_read_all(handle: &mut NSTDStdin, buffer: &mut NSTDVec) -> NSTDIOResult {
     #[cfg(not(unix))]
-    {
-        let (err, r) = crate::io::stdio::read_all(handle, buffer);
-        *read = r;
-        err
-    }
+    return crate::io::stdio::read_all(handle, buffer);
     #[cfg(unix)]
-    {
-        // SAFETY: `handle` owns the file descriptor.
-        unsafe {
-            let (err, r) = crate::os::unix::io::stdio::read_all(handle.lock().as_raw_fd(), buffer);
-            *read = r;
-            err.into()
-        }
+    // SAFETY: `handle` owns the file descriptor.
+    unsafe {
+        crate::os::unix::io::stdio::read_all(handle.lock().as_raw_fd(), buffer).into()
     }
 }
 
@@ -123,7 +100,7 @@ pub fn nstd_io_stdin_read_all(
 /// # Note
 ///
 /// If extending the buffer fails, an error code of `NSTD_IO_ERROR_OUT_OF_MEMORY` will be returned.
-/// This does not mean `read` will return as 0 in this case.
+/// This does not mean there were no bytes read from `handle` in this case.
 ///
 /// # Parameters:
 ///
@@ -131,11 +108,10 @@ pub fn nstd_io_stdin_read_all(
 ///
 /// - `NSTDString *buffer` - The buffer to be extended with data from stdin.
 ///
-/// - `NSTDUInt *read` - Returns as the number of bytes read from stdin.
-///
 /// # Returns
 ///
-/// `NSTDIOError errc` - The I/O operation error code.
+/// `NSTDIOResult read` - The number of bytes read from `handle` on success, or the I/O operation
+/// error code on failure.
 ///
 /// # Panics
 ///
@@ -145,23 +121,13 @@ pub fn nstd_io_stdin_read_all(
 pub fn nstd_io_stdin_read_to_string(
     handle: &mut NSTDStdin,
     buffer: &mut NSTDString,
-    read: &mut NSTDUInt,
-) -> NSTDIOError {
+) -> NSTDIOResult {
     #[cfg(not(unix))]
-    {
-        let (err, r) = crate::io::stdio::read_to_string(handle, buffer);
-        *read = r;
-        err
-    }
+    return crate::io::stdio::read_to_string(handle, buffer);
     #[cfg(unix)]
-    {
-        // SAFETY: `handle` owns the file descriptor.
-        unsafe {
-            let handle = handle.lock();
-            let (err, r) = crate::os::unix::io::stdio::read_to_string(handle.as_raw_fd(), buffer);
-            *read = r;
-            err.into()
-        }
+    // SAFETY: `handle` owns the file descriptor.
+    unsafe {
+        crate::os::unix::io::stdio::read_to_string(handle.lock().as_raw_fd(), buffer).into()
     }
 }